// src/upload/csv-row.dto.ts
export class CsvRowDto {
  CODIGO_CAR_FEDERAL: string;
  NUMEROESTADUAL: string;
  NOMEPROPRIEDADE: string;
  SITUACAO: string;
  MUNICIPIO_CODIGO: string;
  GEOMETRY: string;
  NOMESPROPRIETARIOS: string;
  MODULOS_FISCAIS: string;
  PROTOCOLO: string;
  ATIVIDADE: string;
  CRUZA_TERRA_INDIGENA: string;
  CRUZA_AREA_EMBARGADA: string;
}